Move fixtures to file they are used in
