+++ tests/core/report/output_test.py
@@ -3,6 +3,7 @@ from copy import deepcopy
 from typing import List
 
 import mock
+import pytest
 
 from detect_secrets.core import audit
 from detect_secrets.core.color import AnsiColor
@@ -16,6 +17,46 @@ from detect_secrets.core.report.output import print_summary
 from detect_secrets.core.report.output import print_table_report
 from testing.baseline import baseline
 from testing.baseline import baseline_filename
+
+
+@pytest.fixture
+def live_secrets_fixture():
+    live_secrets: List[ReportedSecret] = [
+        {
+            'failed_condition': ReportSecretType.LIVE.value,
+            'filename': baseline_filename,
+            'line': 90,
+            'type': 'Private key',
+        },
+    ]
+    return live_secrets
+
+
+@pytest.fixture
+def unaudited_secrets_fixture():
+    unaudited_secrets: List[ReportedSecret] = [
+        {
+            'failed_condition': ReportSecretType.UNAUDITED.value,
+            'filename': baseline_filename,
+            'line': 120,
+            'type': 'Hex High Entropy String',
+        },
+    ]
+    return unaudited_secrets
+
+
+@pytest.fixture
+def audited_real_secrets_fixture():
+    audited_real_secrets: List[ReportedSecret] = [
+        {
+            'failed_condition': ReportSecretType.AUDITED_REAL.value,
+            'filename': baseline_filename,
+            'line': 60,
+            'type': 'Hex High Entropy String',
+        },
+    ]
+
+    return audited_real_secrets
 
 
 class TestReportOutput:
